Add URLUtils class for checking URL status and format
A URLUtils class has been created in a new file 'url.py' under 'emote/utils'. The class provides features to send head request and check if a given URL is reachable, as well as checking if the URL has an allowed format. With this, we improve handling and validation of URLs within the application.