Add URL validation to emote request process
A URL validation feature has been added to the emote creation process. The system now checks for a valid URL before making a request. If the URL is found to be invalid, the function returns an error message. Debug messages have also been added to log the URL being accessed and the response status code.