Refactor URL blacklist check and emote naming rule
The 'is_url_blacklisted' function in 'emote/utils/url.py' has been renamed to 'blacklisted_url'. This changes were reflected in different parts of the code to maintain consistency. Additionally, a new 'alphanumeric_name' function was added to check if emote names contain only alphanumeric characters. Previously, this check was performed inline in 'emote/slash_commands.py', but the refactoring allows the rule to be used more generally in the application.